+++ common/src/commonMain/kotlin/com/projeto/nutrilog/domain/usecase/ClearUserUseCase.kt
@@ -1,9 +1,14 @@
 package com.projeto.nutrilog.domain.usecase
 
 import com.projeto.nutrilog.domain.repository.UserRepository
+import com.projeto.nutrilog.domain.repository.ProgressRepository
 
-class ClearUserUseCase(private val repository: UserRepository) {
+class ClearUserUseCase(
+    private val userRepository: UserRepository,
+    private val progressRepository: ProgressRepository
+) {
     suspend operator fun invoke() {
-        repository.clearUser()
+        userRepository.clearUser()
+        progressRepository.clearAllProgress()
     }
 }