+++ common/src/iosMain/kotlin/com/projeto/nutrilog/utils/DateUtils.ios.kt
@@ -1,0 +1,11 @@
+package com.projeto.nutrilog.utils
+
+import platform.Foundation.NSDate
+import platform.Foundation.NSDateFormatter
+
+actual fun getTodayDateString(): String {
+    val formatter = NSDateFormatter().apply {
+        dateFormat = "yyyy-MM-dd"
+    }
+    return formatter.stringFromDate(NSDate())
+}